fixed resource types

+++ src/main/java/com/github/CodeNekomancer/OADA_Backend/model/ResourceComp/ResourceComp.java
@@ -16,9 +16,9 @@ public class ResourceComp {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long resourceComp_ID;
 
-    private Integer metal;
-    private Integer crystal;
-    private Integer duty;
+    private Long metal;
+    private Long crystal;
+    private Long duty;
     private Integer MO;
     private Integer energy;
